Finished 3p, started 4p
Finished up a reward shaping version of 3-player games with Marquise, Eyrie, and Alliance.
A simple reward function is to give the model a small reward for every victory point gained.
Attempted to make Dominance Cards as "viable" as points, but rewarding only when they end up in a winning position, and punishing them when they have a dominance in a bad spot to win with it.

Started implementing the Vagabond. 4p, here we come!