More VB Implementation
Continued work on the VB's action-handling methods; the daylight phase contains like 90% of the Vagabond's choices, so I've gotten a good portion of that done so far.

Kind of dreading implementing the battle functionality for the VB, not sure how complicated it will really be. I want to end up rewriting the gigantic, ugly battle function I'm currently using in the process to make it more like the system I have to handle actions in general:
- One method reads the current state of the game and returns a list of actions the current player has to choose from
- A separate method takes an action and executes it, changing the game state and current player as necessary.