4P Env V2 - Condensed Obs
Made a copy of the 4 Player environment. The only thing I really changed was all of the observation representations of each part of the game. I changed the observation Box to be in the range [-1,4], allowing for many items' values to be exact inputs (hand size up to 4, VB item counts, etc).

Additionally, I made a large number of observations normalized in [0,1] or [0,4], the effects of which I will have to test. This saved a ton of size and cut down on the complexity of the observation space. I also changed the seen hands memory size to 2 (saves the current turn and the previous one).

In all, the observation space alone (no action space added) was cut down to about 1/3 its previous size, while hopefully conveying the same or better information.

A few last changes:
- Fixed a bug where NOBODY could see who the chosen VB coalition partner is. Oops.
- The VB is given a number that hints at how close they are to going over their bag limit
- The VB is given the exact number of quests of a suit completed so far
- The Eyrie are now told at all times the suit counts in each part of their decree. Before, I just told them what cards were where and the remaining decree requirements during their turn, but they could not see this while adding to their decree. Things have been really unfair for them...
- The Alliance is given an exact count of their supporters, until just being told "5+ supporters" if it's high enough
- The ID is explicitly given for the next player in turn order
- The VB "aids done this turn" is now a ratio of how close they are to improving the given relationship