final changes before 4pv2 testing
Made a few changes to the observation before starting training:
- Explicitly tells the number of empty building slots in a clearing
- Explicitly tells who is the ruler of a clearing if there is one
- Tells a number from 0 to about 3.5 that represents the number of actions played in this game so far

All rewards are exactly the same as in the first 4p implementation except for the reward for winning, which was increased from 8 and 12 to 50 and 60 (points vs dom win compensation). To adjust for this, all rewards are scaled appropriately smaller. The average win seems to be around +0.8, making the average loss at about -0.26.67. However, the variation in winning and losing is much smaller; losers seem to always be negative, at best close to -0.05 if they did really well.

Lastly, I will need to do lots of testing with both hyperparameters, but also model sizes.