final updates to base 4p V2
For reasons I am not completely sure of, the model was completely overfitting at every step when the inputs were above 1. Thus, I normalized every input to be in the range of [-1,1].
Additionally, I changed the rewards back to about what they were before, with wins giving 10 and 12 points for normal / domination victories.
Other changes:
- Max actions in one game reduced to 1000
- Added more debugging so I could see if the VB / Eyrie were actually winning during testing
- Added an array to tell the model which items had been found in the ruins already
- Made it so a VB with no items left to damage cannot be attacked