testing for single player game learning
Added environments for testing if the SIMPLE PPO system can learn singleplayer games. After a bit of testing, it definitely can, though the process is not quite as smooth.
Since we want to update the opponents as well, many separate training sessions will need to be managed, unless I come up with some script that can automate it?
- Tested "Singleplayer TicTacToe", which is an environment consisting of just playing a random tictactoe player. The default training regiment made a model that could win >90% of the time in about 6 minutes.
- Started work and testing on a simple 2 player, asymmetric game I call Proxy, which is kind of like tag.