per-faction models start
After the success of the 'proxychaser' toy game for models to learn one role in an asymmetric game at a time, I am now converting the previous ROOT implementation into a slightly different environment that can handle different players playing it at the same time. In doing this, I am also making other changes to the action mechanics that I think will help the models improve.

So far, I have updated the Marquise and the Eyrie.

- Each faction will now have its own unique action numbers, although I will reserve a number of them to be the same across all factions for actions that every one can do, such as discarding a card or using cards in battle.
- A 'Move' is now composed of two separate actions: Choosing the start and end clearings of the move, and then choosing how many warriors should make the previously chosen move. I hope this will be a little easier to learn, since this reduces the number of actions to understand by about 3,300. Models will also be much smaller in terms of the number of weights to adjust because of this.
- A few other actions involving combinations of moves were split into a sequence of choices, such as Overworking sawmills, adding to the decree, and taking available dominance cards